Switched to using static functions for localization converter.
