+++ src/AmbientSounds.Uwp/AmbientSounds.Uwp.csproj
@@ -130,8 +130,7 @@
     <Compile Include="Controls\SleepTimerControl.xaml.cs">
       <DependentUpon>SleepTimerControl.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Converters\PlayButtonAutomationNameConverter.cs" />
-    <Compile Include="Converters\SoundNameConverter.cs" />
+    <Compile Include="Converters\LocalizationConverter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\LocalSettings.cs" />
     <Compile Include="Services\MediaPlayerService.cs" />
