+++ src/AmbientSounds.Uwp/Controls/PlayerControl.xaml
@@ -2,6 +2,7 @@
     x:Class="AmbientSounds.Controls.PlayerControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:converters="using:AmbientSounds.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:local="using:AmbientSounds.Controls"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
@@ -14,10 +15,10 @@
             <Button
                 Width="60"
                 Height="60"
-                AutomationProperties.Name="{x:Bind ViewModel.IsPaused, Mode=OneWay, Converter={StaticResource PlayButtonAutomationNameConverter}}"
+                AutomationProperties.Name="{x:Bind converters:LocalizationConverter.ConvertPlayerButtonState(ViewModel.IsPaused), Mode=OneWay}"
                 Command="{x:Bind ViewModel.TogglePlayStateCommand}"
                 CornerRadius="30"
-                ToolTipService.ToolTip="{x:Bind ViewModel.IsPaused, Mode=OneWay, Converter={StaticResource PlayButtonAutomationNameConverter}}">
+                ToolTipService.ToolTip="{x:Bind converters:LocalizationConverter.ConvertPlayerButtonState(ViewModel.IsPaused), Mode=OneWay}">
                 <Grid>
                     <!--  play  -->
                     <FontIcon
@@ -36,7 +37,7 @@
                 Margin="20,0,0,0"
                 VerticalAlignment="Center"
                 Background="Transparent">
-                <TextBlock Margin="0,8,0,0" Text="{x:Bind ViewModel.SoundName, Mode=OneWay, Converter={StaticResource SoundNameConverter}}" />
+                <TextBlock Margin="0,8,0,0" Text="{x:Bind converters:LocalizationConverter.ConvertSoundName(ViewModel.SoundName), Mode=OneWay}" />
                 <Slider
                     Width="120"
                     Margin="0,4,0,0"
