+++ src/AmbientSounds.Uwp/Views/MainPage.xaml
@@ -4,6 +4,7 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:animations="using:Microsoft.Toolkit.Uwp.UI.Animations"
     xmlns:controls="using:AmbientSounds.Controls"
+    xmlns:converters="using:AmbientSounds.Converters"
     xmlns:core="using:Microsoft.Xaml.Interactions.Core"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:extensions="using:Microsoft.Toolkit.Uwp.UI.Extensions"
@@ -74,7 +75,7 @@
                         <GridViewItem
                             Width="300"
                             Height="180"
-                            AutomationProperties.Name="{x:Bind AutomationName, Mode=OneWay}">
+                            AutomationProperties.Name="{x:Bind converters:LocalizationConverter.ConvertGridViewSoundName(IsCurrentlyPlaying, Unplayable, Name), Mode=OneWay}">
                             <GridViewItem.Resources>
                                 <!--  These two removes the border on hover and on pressed  -->
                                 <SolidColorBrush x:Key="SystemControlHighlightListLowBrush" Color="Transparent" />
@@ -108,7 +109,7 @@
                                         FontSize="36"
                                         FontWeight="SemiBold"
                                         Foreground="White"
-                                        Text="{x:Bind Name, Converter={StaticResource SoundNameConverter}}"
+                                        Text="{x:Bind converters:LocalizationConverter.ConvertSoundName(Name)}"
                                         TextWrapping="Wrap" />
                                 </Grid>
                             </Grid>
