+++ src/AmbientSounds/ViewModels/SoundViewModel.cs
@@ -23,7 +23,6 @@ namespace AmbientSounds.ViewModels
             _playerService.PlaybackStateChanged += PlayerService_PlaybackStateChanged;
         }
 
-
         /// <summary>
         /// The sound's attribution.
         /// </summary>
@@ -40,14 +39,14 @@ namespace AmbientSounds.ViewModels
         public string ImagePath => _sound.ImagePath;
 
         /// <summary>
-        /// Determines the name that represents the purpose of this item most accurately.
+        /// Returns true if the sound cannot be played.
         /// </summary>
-        /// Idea: First check if we have a file to play, if not, indicate that our sound is not playable.
-        /// After that, check if service is playing and the current item is our sound.
-        /// If so, we can be paused, otherwise the sound can be played.
-        public string AutomationName => string.IsNullOrEmpty(_sound.FilePath) ? Name + ", not playable" :
-            (_playerService.PlaybackState == MediaPlaybackState.Playing
-                && _playerService.Current == _sound ? "Pause " + Name : "Play " + Name);
+        public bool Unplayable => string.IsNullOrWhiteSpace(_sound.FilePath);
+
+        /// <summary>
+        /// Returns true if the sound is currently playing.
+        /// </summary>
+        public bool IsCurrentlyPlaying => _playerService.PlaybackState == MediaPlaybackState.Playing && _playerService.Current == _sound;
 
         /// <summary>
         /// Loads this sound into the player and plays it.
@@ -59,7 +58,7 @@ namespace AmbientSounds.ViewModels
 
         private void PlayerService_PlaybackStateChanged(object sender, MediaPlaybackState e)
         {
-            OnPropertyChanged(nameof(AutomationName));
+            OnPropertyChanged(nameof(IsCurrentlyPlaying));
         }
     }
 }
